requirements.txt
Flask
pymongo>=4.6
python-dotenv>=1.0
